Move title to shared carousel
