+++ src/components/contentContainer/newsCarousel.tsx
@@ -4,12 +4,7 @@ import styled, { css } from 'styled-components';
 
 import { NewsCarouselInformationFragment } from '../../../types/graphql-types';
 import Carousel from '../elements/carousel';
-import Title from '../elements/title';
-import {
-  applyMediaQueryMd,
-  applyMediaQueryLg,
-  contentMargin,
-} from '../../style/dimensions';
+import { applyMediaQueryMd } from '../../style/dimensions';
 import NewsSlide, {
   slideMarginRight,
   slideWidth,
@@ -27,18 +22,6 @@ const Container = styled.div`
   `)}
 `;
 
-const CarouselTitle = styled(Title)`
-  padding: 0 ${contentMargin.sm} 15px;
-
-  ${applyMediaQueryMd(css`
-    padding: 0 ${contentMargin.md} 34px;
-  `)}
-
-  ${applyMediaQueryLg(css`
-    padding: 0 ${contentMargin.lg} 34px;
-  `)}
-`;
-
 const NewsCarousel: FC<Props> = ({ data }) => {
   const slides = useMemo(
     () =>
@@ -50,11 +33,11 @@ const NewsCarousel: FC<Props> = ({ data }) => {
 
   return (
     <Container>
-      <CarouselTitle type="h3" title={data.title!} />
       <Carousel
         slides={slides}
         slideMarginRight={slideMarginRight}
         slideWidth={slideWidth}
+        title={data.title}
       />
     </Container>
   );
